+++ packages/beagle-react/src/components/BeagleListView/styled.ts
@@ -19,9 +19,13 @@ import { Direction } from 'common/models'
 
 interface StyledListViewInterface {
   direction: Direction,
+  useParentScroll?: boolean,
 }
 
 export const StyledListView = styled.div<StyledListViewInterface>`
   display: flex;
-  flex-direction: ${({ direction }) => direction === 'VERTICAL' ? 'column' : 'row'}
+  flex-direction: ${({ direction }) => direction === 'VERTICAL' ? 'column' : 'row'};
+  overflow: ${({ useParentScroll }) => useParentScroll ? 'inherit' : 'auto'};
+  width: ${({ direction }) => direction === 'HORIZONTAL' ? '100%' : 'auto'};
+  height: ${({ direction }) => direction === 'VERTICAL' ? '100%' : 'auto'};
 `
