+++ packages/beagle-react/src/components/index.ts
@@ -21,7 +21,6 @@ import BeagleLoading from './BeagleLoading'
 import BeagleError from './BeagleError'
 import BeagleImage from './BeagleImage'
 import BeagleListView from './BeagleListView'
-import BeagleFutureListView from './BeagleFutureListView'
 import BeagleLazy from './BeagleLazy'
 import BeagleTouchable from './BeagleTouchable'
 import BeagleSimpleForm from './BeagleSimpleForm'
@@ -45,7 +44,6 @@ const beagleDefaultComponents = {
   'beagle:button': BeagleButton,
   'beagle:text': BeagleText,
   'beagle:listview': BeagleListView,
-  'beagle:futurelistview': BeagleFutureListView,
   'beagle:container': BeagleContainer,
   'beagle:screencomponent': BeagleScreen,
   'beagle:pageview': BeaglePageView,
